+++ packages/orq-rc/RELEASES.md
@@ -2179,3 +2179,13 @@ Based on:
 - [typescript v3.8.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.24 - packages/orq-rc
+
+## 2025-06-13 10:25:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.561.0 (2.628.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.25 - packages/orq-rc